Inline table primitives
