+++ web/.storybook/docs/ThemeTokens/Color.tsx
@@ -211,7 +211,7 @@ function InteractionStatesSection({ ctx }: { ctx: TokenContext }) {
           state="Hover · row"
           classes="hover:bg-muted/50"
           tokens="--muted at 50%"
-          seenIn="TableRow (ui/table.tsx)"
+          seenIn="DataTable / SimpleDataTable"
           sample={
             <div
               className="flex flex-col rounded-md border text-xs"
@@ -231,7 +231,7 @@ function InteractionStatesSection({ ctx }: { ctx: TokenContext }) {
           state="Selected · row"
           classes="data-[state=selected]:bg-muted"
           tokens="--muted"
-          seenIn="TableRow (ui/table.tsx)"
+          seenIn="DataTable / SimpleDataTable"
           sample={
             <div
               className="flex flex-col rounded-md border text-xs"
